Change api_version to separate request_version/response_version.
This is important for apps that use HTTP content negotiation for 
versioning.  The response version would be selected based on the
Content-Type of the response (which the server might select based
on the Accept request header).  The request version would be
selected based on the Content-Type of the request.

They need to be separately selectable, hence, separate configs.