Fix DynamicStruct so that it allows a `sort` param.
`sort` params were broken due to the fact that we
are using a Hashie::Mash as our DynamicStruct. Hash
has a sort method, so it meant that a `sort` entry
in the hash was not used for `mash.sort`.

I don't remember the full history of why I switched to
a Hashie::Mash in 8f6d2f4e but this is one downside:
more collisions with method names. We may want to
reconsider that decision at some point, but this solves
the problem for now.